requirements/base.txt
django==1.9.12
django-extensions==1.7.5
django-waffle==0.11.1
djangorestframework==3.4.7
edx-auth-backends==0.5.2
edx-rest-api-client==1.6.0
python-dateutil==2.5.3
pytz==2016.10
wagtail==1.7
git+https://github.com/moorinteractive/wagtail-bakery.git#egg=wagtail-bakery
wagtailmenus==2.0.3
